Refactor lrowaccal for testability and add a radiance units label to the output cube (#4940)
* Add radiance units label in lrowaccal output cube

* Refactor lrowaccal app to be callable for testing and add radiance units to output PVL

* Update formatting and NULL initialization

* Add NULL assignments after deletions

* Add CubeAttributeOutput to output cube initialization

* Add functional test for lrowaccal units label addition and accompanying test cube

* Refactor lrowaccal to remove global variables and convert helper functions to lambdas with captures

* Update history section of lrowaccal XML file with refactoring and output cube units label changes

* Remove old lrowaccal Makefile tests

* Add lrowaccal functional test to ensure that the radiance units label is not inserted for RadiometricType IOF

* Add Cordell Michaud to the .zenodo.json document

* Add lrowaccal changes to CHANGELOG.md

* Clean up lrowaccal includes and leftover comment

* Re-add lrowaccal Makefile tests

* Re-add disclaimer to top of lrowaccal's main.cpp file

* Add input attributes retrieval to lrowaccal.cpp

* Fix lrowaccal functional test error messages and add PVl group and existence checks

* Move lrowaccal units change message to Changed sub-header in CHANGELOG.md

* Simplify lrowaccal functional test assertions
