+++ .zenodo.json
@@ -305,6 +305,10 @@
       "name": "Mehlman, Bob"
     },
     {
+      "affiliation": "Arizona State University",
+      "name": "Michaud, Cordell"
+    },
+    {
       "name": "Milazzo, Moses"
     },
     {
